Nidal

+++ src/parent.c
@@ -1,18 +1,23 @@
 #include "parent.h"
 
+pid_t children[NUM_CHILDREN];
+bool confirmed[NUM_CHILDREN - 1] = {false};
+float numbers[NUM_CHILDREN - 1] = {0.0};
 int main(int argc, char *argv[]) {
-	handler_setup(SIGUSR1, &ready_to_start);
-	pid_t *children = create_children(NUM_CHILDREN);
+	handler_setup(SIGUSR1, &child_confirmations);
+	create_children(NUM_CHILDREN);
+	sleep(1); // Wait for children to finish setting up their signal handlers
 	write_range("./txt/range.txt", 1, 100);
-	while(ready_counter < 5) pause(); // Wait until all children are ready (Note: without this line, the parent will send the signal before the children finish setting up their signal handlers)
-	for (int j=0; j < NUM_CHILDREN; ++j) kill(children[j], SIGUSR1); // Send SIGUSR1 to all children (Start signal)
+	for (int j=0; j < NUM_CHILDREN - 1; ++j) kill(children[j], SIGUSR1); // Send SIGUSR1 to all children (Start signal)
+	kill(children[NUM_CHILDREN - 1], SIGPIPE); // Send SIGUSR2 (PIPE Child)
+	while(ready_counter < NUM_CHILDREN - 1) pause(); // Wait for all children to send SIGUSR1 (Confirm signal)
+	get_numbers();
 	while(wait(NULL) > 0); // Wait for all children to finish
 	return 0;
 }
 
-pid_t *create_children()
+void create_children()
 {
-	static pid_t children[NUM_CHILDREN]; // static so that it doesn't get destroyed when the function returns
 	for (int i = 0; i < NUM_CHILDREN; i++)
 	{
 		pid_t pid = fork();
@@ -33,39 +38,26 @@ pid_t *create_children()
 			perror("fork");
 			exit(1);
 		}
-		pause(); // Wait for child to finish setting up its signal handler
 	}
-	return children;
 }
 
-void write_range(char *filename, int min, int max)
-{
-	// write two comma-separated integer values  in a file called range.txt
-	FILE *fp = fopen(filename, "w");
-	if (fp == NULL)
-	{
-		perror("fopen");
-		exit(1);
+void child_confirmations(int signo, siginfo_t *info, void *context) {
+	// printf("[Parent]\n");
+    // printf("Received signal %d\n", signo);
+    // printf("Sender PID: %d\n", info->si_pid);
+	int index = getIndex(children, info->si_pid, NUM_CHILDREN); //info->si_pid - children[0];
+	if (!confirmed[index]) {
+		ready_counter += 1;
+		kill(info->si_pid, SIGUSR2); // Send SIGUSR2 to child (Confirm signal)
+		confirmed[index] = true;
 	}
-	fprintf(fp, "%d,%d", min, max);
-	fclose(fp);
 }
 
-void ready_to_start(int signo, siginfo_t *info, void *context) {
-	printf("[Parent]\n");
-    printf("Received signal %d\n", signo);
-    printf("Sender PID: %d\n", info->si_pid);
-	kill(info->si_pid, SIGUSR2); // Send SIGUSR1 to child (Ready signal
-	ready_counter += 1;
-}
-
-void handler_setup(int sig, void (*handler)(int, siginfo_t *, void *)) {
-	struct sigaction sa;
-	sa.sa_handler = (void (*)(int)) handler;
-	sigemptyset(&sa.sa_mask);
-	sa.sa_flags = SA_SIGINFO;
-	if (sigaction(sig, &sa, NULL) == -1) {
-		perror("sigaction");
-		exit(1);
+void get_numbers() {
+	for (int i = 0; i < NUM_CHILDREN - 1; i++) {
+		char filename[20];
+		sprintf(filename, "./txt/%d.txt", children[i]);
+		numbers[i] = read_number(filename);
+		printf("Child %d: %f\n", children[i], numbers[i]);
 	}
 }
